Moved web server logic back to server
